A guardrail at its own scan bound will not call itself clear
Review found the one failure this reading exists to prevent, in the reading
itself.

WaitingScanCap bounds the shared statement, so all five reads clip at the
same 200. On a queue already at the cap the strict read and every relaxed
read return 200, all four differences are zero, and every figure honestly
says "nothing hidden" — over the installation most likely to be hiding work.
No assertion anywhere would fail, which is the shape of defect the file's own
comments call the one a guardrail must never have.

So a truncated read is now reported, and `clear` is false whenever it is set.
That is not a claim that work IS hidden; it is a refusal to claim the
opposite over a question the scan stopped before settling.

The integration test seeds one row PAST the cap rather than mocking the flag:
the clipping belongs to the LIMIT, and a fake would prove only that a boolean
can be assigned. It first checks the fixture actually reached the bound, so a
future change to the cap fails loudly instead of quietly testing nothing.

attentionlanesseam.go crossed the 500-line cap with the extra projection, so
the who-is-waiting binding moved to attentionwaitingseam.go. It is the lane
read BESIDE the assembled day rather than as one of its fourteen — it carries
its own truncation answer, its own machine-sender filter and now a second
question about the same query, while its siblings in that file are each a
single pass-through.

Nothing else from the review: the two new SQL holes take compile-time
constants only, both relaxations sit inside their own parentheses so FALSE OR
X is exactly X, and the zero-uuid reader relaxes only the reader-state
subquery it appears in.
