+++ backend/internal/contracts/api_gen.go
@@ -20356,9 +20356,9 @@ type HiddenBacklog struct {
 	// AsOf The instant every figure below was read at.
 	AsOf time.Time `json:"as_of"`
 
-	// Clear True when nothing is being held back. The guardrail's target, sent as its own
-	// field because a number only ever read beside other numbers becomes decoration —
-	// this is what a check asserts on.
+	// Clear True when nothing is being held back AND the reads were complete. The
+	// guardrail's target, sent as its own field because a number only ever read beside
+	// other numbers becomes decoration — this is what a check asserts on.
 	Clear bool `json:"clear"`
 
 	// NotSales Work somebody judged to be no business of the queue's. SOMEBODY'S CHOICE, and
@@ -20387,6 +20387,19 @@ type HiddenBacklog struct {
 	// rather than as bare volumes — three hidden against four shown is a broken
 	// queue, and three against three hundred is a rep tidying up.
 	Shown int `json:"shown"`
+
+	// Truncated True when a read stopped at its own scan bound, which makes every figure above
+	// it a FLOOR rather than a count.
+	//
+	// The bound is on the shared statement, so the strict read and every relaxed read
+	// clip at the same number. On a queue already at the cap all five return it, every
+	// difference is zero, and a guardrail without this flag would report a clear
+	// backlog over the installation most likely to be hiding work.
+	//
+	// `clear` is false whenever this is true. That is not a claim that work IS hidden;
+	// it is a refusal to claim the opposite over a question the scan stopped before
+	// settling.
+	Truncated bool `json:"truncated"`
 
 	// Unlinked Inbound mail that qualifies in every other way and is attached to no record the
 	// workspace sells to. ALSO NOBODY'S CHOICE, and genuinely ambiguous: usually it
